chore: Code maintenance
No functional changes

* Extract `GitExtLinkFormat` and rename to `ExternalLinkFormat`
* Rename `ExternalLinkFormat.ToGitExtLink` to `Apply` to increase API
readability
* Reorder ExternalLinkDefinition members to make it easier to see what
forms the serialisation contract and what represents the behavioural aspect
of the existing implementation.
